Speed up (?) workflow
Attempt to possibly speed up workflow for some users. Remove full choochoo install to check if an instructor/admin/student. Instead test if i) command starts with choo-choo ii) issue-commenter is issue-owner. No job for raising issues (which will only be done by admin/instructor), just a general check if it is an admin/instructor thread that that person is an admin/instructor.